requirements.txt
numpy
pandas
-e .
pymysql
python-dotenv
mysql-connector-python
scikit-learn
seaborn
matplotlib
catboost
xgboost